[190. Reverse Bits][Accepted]committed by LeetCode Extension

+++ 190-Reverse-Bits/solution.java
@@ -1,8 +1,8 @@
 public class Solution {
     // you need treat n as an unsigned value
     public int reverseBits(int n) {
-        long ans = 0;
-        long num = (long)n;
+        int ans = 0;
+        int num = (int)n;
         for(int i = 0; i < 32; i++){
             ans *= 2;
             ans += num%2;
